added docs, context
